Refactor compiler to be component
